fix - fix users

+++ api/src/users/users.service.ts
@@ -3,6 +3,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
 import { User } from './user.entity';
+import * as bcrypt from 'bcryptjs';
+import { HttpException, HttpStatus } from '@nestjs/common';
 
 @Injectable()
 export class UsersService {
@@ -11,13 +13,28 @@ export class UsersService {
     private readonly usersRepository: Repository<User>,
   ) {}
 
-  create(createUserDto: CreateUserDto): Promise<User> {
+  async create(createUserDto: CreateUserDto): Promise<User> {
+    const existingUser = await this.usersRepository.findOneBy({
+      userName: createUserDto.userName,
+    });
+
+    if (existingUser) {
+      throw new HttpException(
+        'Username already exists',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     const user = new User();
     user.firstName = createUserDto.firstName;
     user.lastName = createUserDto.lastName;
     user.userName = createUserDto.userName;
-    user.password = createUserDto.password;
+
+    const salt = await bcrypt.genSalt(10);
+    user.password = await bcrypt.hash(createUserDto.password, salt);
+
     user.isActive = createUserDto.isActive;
+
     return this.usersRepository.save(user);
   }
 
@@ -30,6 +47,12 @@ export class UsersService {
   }
 
   async remove(userName: string): Promise<void> {
+    const user = await this.usersRepository.findOneBy({ userName });
+
+    if (!user) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
     await this.usersRepository.delete(userName);
   }
 }
